CashBook_Backend/requirements.txt
fastapi==0.118.0
pydantic==2.11.10
pydantic_core==2.33.2
pymongo==4.15.2
python-dotenv==1.1.1
uvicorn==0.37.0
pandas==2.3.3
python-multipart==0.0.20
openpyxl==3.1.5
xlrd==2.0.2
email-validator==2.2.0
dnspython==2.8.0
numpy==2.3.4
six==1.17.0
et-xmlfile==2.0.0
tzdata==2025.2
pytz==2025.2
starlette==0.48.0
xlsxwriter==3.2.0
fpdf2==2.8.2
passlib==1.7.4
bcrypt==4.2.0
